move input panel to util
