+++ src/inspect_ai/approval/_human/panel.py
@@ -9,10 +9,10 @@ from textual.reactive import reactive
 from textual.widgets import Button, Static
 from typing_extensions import override
 
-from inspect_ai._display.core.input import InputPanel
 from inspect_ai._util.registry import registry_unqualified_name
 from inspect_ai.solver._task_state import TaskState
 from inspect_ai.tool._tool_call import ToolCall, ToolCallView
+from inspect_ai.util._panel import InputPanel, input_panel
 
 from .._approval import Approval, ApprovalDecision
 from .manager import ApprovalRequest, PendingApprovalRequest, human_approval_manager
@@ -34,10 +34,8 @@ async def panel_approval(
     state: TaskState | None,
     choices: list[ApprovalDecision],
 ) -> Approval:
-    from inspect_ai._display.core.active import task_screen
-
     # ensure the approvals panel is shown
-    await task_screen().input_panel(PANEL_TITLE, ApprovalInputPanel)
+    await input_panel(PANEL_TITLE, ApprovalInputPanel)
 
     # submit to human approval manager (will be picked up by panel)
     approvals = human_approval_manager()
